spotific/requirements/requirements.txt
django==4
django-crispy-forms
oauth2
Pillow
psycopg2
spotipy
numerize
pandas
matplotlib
tweepy
wikipedia